+++ example/lib/example.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:material_floating_search_bar_2/material_floating_search_bar_2.dart';
+import 'package:material_floating_search_bar_plus/material_floating_search_bar_plus.dart';
 
 class Example extends StatefulWidget {
   const Example({super.key});